+++ tasks/makovskiy_i_gauss_filter_vert/seq/src/ops_seq.cpp
@@ -40,7 +40,7 @@ bool GaussFilterSEQ::RunImpl() {
         for (int k_col = -1; k_col <= 1; ++k_col) {
           sum += GetPixel(input, col + k_col, row + k_row, width, height) *
                  // NOLINTNEXTLINE(cppcoreguidelines-pro-bounds-constant-array-index)
-                 kernel[static_cast<size_t>(((k_row + 1) * 3) + (k_col + 1))];
+                 kernel[(static_cast<size_t>(k_row + 1) * 3) + static_cast<size_t>(k_col + 1)];
         }
       }
       output[(static_cast<size_t>(row) * width) + col] = sum / kernel_sum;
